added a flag to prevent memmgr.free sending a freed resources message when doing resource availability tests to avoid false flagging of available resources

+++ qoala/sim/memmgr.py
@@ -123,7 +123,7 @@ class MemoryManager(Protocol):
 
         return self.allocate(pid, virt_id)
 
-    def free(self, pid: int, virt_id: int) -> None:
+    def free(self, pid: int, virt_id: int, test: bool = False) -> None:
         vmap = self._process_mappings[pid]
         # Check if the virtual ID is in the unit module
         assert virt_id in vmap.unit_module.get_all_qubit_ids()
@@ -141,7 +141,8 @@ class MemoryManager(Protocol):
         self._qdevice.set_mem_pos_in_use(phys_id, False)
 
         # send a signal for components that may be blocked on resources
-        self.send_signal(SIGNAL_MEMORY_FREED)
+        if not test:
+            self.send_signal(SIGNAL_MEMORY_FREED)
 
     def get_unmapped_non_comm_qubit(self, pid: int) -> int:
         """returns virt ID"""
